refactor(loaders): move loaders into separate directory
