+++ src/loaders/text.js
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 Adobe. All rights reserved.
+ * Copyright 2019 Adobe. All rights reserved.
  * This file is licensed to you under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License. You may obtain a copy
  * of the License at http://www.apache.org/licenses/LICENSE-2.0
@@ -9,19 +9,12 @@
  * OF ANY KIND, either express or implied. See the License for the specific language
  * governing permissions and limitations under the License.
  */
-/* eslint-disable no-underscore-dangle, no-param-reassign */
-import { nest, cast } from './util.js';
+import { nest, cast } from '../util.js';
 
-function load(str, prefix = '') {
-  const obj = cast(new URLSearchParams(str));
+function load(txt) {
+  const pairs = txt.split('\n').map((l) => l.split('=').slice(0, 2)).filter(([k]) => !!k);
 
-  return nest(Object.entries(obj).reduce((o, [k, v]) => {
-    if (k.startsWith(prefix)) {
-      // eslint-disable-next-line no-param-reassign
-      o[k.replace(prefix, '')] = v;
-    }
-    return o;
-  }, {}));
+  return nest(cast(pairs));
 }
 
 export { load };
